requirements.txt
# SNN Optical Flow - Requirements
# Python 3.8+

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Data Processing
numpy>=1.21.0
scipy>=1.7.0
pillow>=9.0.0
h5py>=3.7.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Logging and Monitoring
tensorboard>=2.10.0
tqdm>=4.64.0

# Configuration
pyyaml>=6.0

# Optional but recommended
# For better performance
einops>=0.6.0

# For data augmentation
albumentations>=1.3.0

# For video processing (if needed)
opencv-python>=4.5.0

# For event camera simulation/processing
# h5py>=3.7.0

# Development tools (optional)
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0

# FPGA deployment tools (install separately based on your FPGA vendor)
# These are just placeholders - actual tools depend on target FPGA
# vivado (Xilinx)
# quartus (Intel/Altera)
# brevitas (for quantization - optional advanced tool)
